refactor(cli): extract reusable terminal selectors
- Isolate generic arrow-key picker primitives from MCP elicitation flow.
- Keep elicitation-specific rendering and validation in the REPL handler.
- Make list selection reusable for future interactive CLI surfaces.
